Solve day 2, part 2
Wrote a bunch of scenarios based on the example in the puzzle only to
realise that the minimums are also the maximums so very little new code
is required to get to the answer.

I'm a little embarrassed by the structure and organisation of these
specs, but they serve the purpose, which is to give me confidence that
the code is doing what it's supposed to.